streamlit_requirements.txt
streamlit==1.22.0
pyOpenSSL==23.1.1
cryptography==40.0.2
